+++ chatter-cloud/src/main/java/com/lens/chatter/service/ProductPhotoService.java
@@ -3,11 +3,8 @@ package com.lens.chatter.service;
 import com.lens.chatter.constant.ErrorConstants;
 import com.lens.chatter.constant.HttpSuccessMessagesConstants;
 import com.lens.chatter.exception.BadRequestException;
-import com.lens.chatter.exception.NotFoundException;
-import com.lens.chatter.model.entity.Product;
 import com.lens.chatter.model.entity.ProductPhoto;
 import com.lens.chatter.repository.ProductPhotoRepository;
-import com.lens.chatter.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,24 +21,14 @@ import java.util.UUID;
 public class ProductPhotoService {
 
     @Autowired
-    private ProductRepository productRepository;
+    private ProductService productService;
 
     @Autowired
     private ProductPhotoRepository repository;
 
     @Transactional
     public String uploadProductPhoto(MultipartFile file, UUID productId) {
-        Product product = productRepository.findById(productId).orElse(null);
-        if (product == null) {
-            throw new NotFoundException(ErrorConstants.PRODUCT_NOT_EXIST);
-        }
-        ProductPhoto photo;
-        if (repository.existsByProductId(productId)) {
-            throw new BadRequestException(ErrorConstants.PRODUCT_PHOTO_ALREADY_EXIST);
-        } else {
-            photo = new ProductPhoto();
-            photo.setProduct(product);
-        }
+        ProductPhoto photo = repository.findProductPhotoByProductId(productId).orElse(new ProductPhoto(productService.fromIdToEntity(productId)));
         try {
             photo.setFile(file.getBytes());
         } catch (IOException e) {
@@ -53,16 +40,19 @@ public class ProductPhotoService {
 
     @Transactional
     public byte[] getPhoto(UUID productId) {
-        ProductPhoto photo = repository.findProductPhotoByProductId(productId);
+        ProductPhoto photo = repository.findProductPhotoByProductId(productId).orElse(null);
         if (photo == null) {
-            throw new BadRequestException(ErrorConstants.PRODUCT_PHOTO_NOT_EXIST);
+            throw new BadRequestException(ErrorConstants.DEFECT_PHOTO_NOT_EXIST);
         }
         return photo.getFile();
     }
 
     @Transactional
     public void deletePhotoByProductId(UUID productId) {
-        repository.deleteProductPhotoByProductId(productId);
+
+        if (repository.existsByProductId(productId)) {
+            repository.deleteProductPhotoByProductId(productId);
+        }
     }
 
     @Transactional
